coin: Skip tests based on DoNotRunTestsButCreateTestWorkItem feature
Some build configurations might have their own instructions for
the test work item, where it doesn't make sense to run the Qt
auto tests.

We can't use DisableTests for this, as that is hard-coded in coin
to skip the entire test work item.

Nor can we use DoNotBuildTests, as the test config might not be
building anything (depends on another build config, that does build
the tests).

Nor can we use DoNotRunTests, as that is hard-coded in coin
to skip the entire work item. We should probably fix that,
but until then let's use a dedicated feature to selectively
skip running the tests while still creating the test work item.

By using this feature we can for example let the documentation
warning configuration build docs and check for new warnings,
without also running Qt auto tests in the same test run.

Change-Id: I108bece654c811e98ba06e07e4e7b06f6d84c8f4
Reviewed-by: Paul Wicking <paul.wicking@qt.io>