fix: Redundant condition in qtextdocument
Q_ASSERT(n || (!n && block_pos == blocks.length()))
contains a redundant conditional check,when n is true,the
conditional judgment must be true,when n is false,whether
the conditional judgment is true or not depends entirely
on block_pos == blocks.length( ), “!n” is not necessary.
eg: 'A || (!A && B)' is equivalent to 'A || B'.

Change-Id: Ia1c2df548d97db92a4fa1089e3f5a79e45800829
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>