a11y: Notify of implicit window title change
Commit 25ab8ada1019adf9e88addd47acb0924831edc5a
implemented sending a QAccessible::NameChanged
event when the window name is used as accessible
name and is changed by calling
QWidget::setWindowTitle.

Besides explicitly setting a new window title that way,
the title can also be changed less explicitly:
If the window title contains a corresponding "[*]"
placeholder, changing the QWidget::windowModified
property also results in a changed window title,
with no QAccessible::NameChanged being sent so
far.

Move the logic introduced in
25ab8ada1019adf9e88addd47acb0924831edc5a
further down into QWidgetPrivate::setWindowTitle_sys
so it covers that case as well.

While at it, also adjust this to skip retrieving
the accessible here after all if accessibility
is not enabled.

This also needs
5178606a98be57f6dc5f3dddc1fa0f1e16c933e5
(for xcb; corresponding qtwayland still pending
in Gerrit) for the event to actually be
sent for the QWidget::setWindowModified
case, because oldAccessibleName
in QWidgetPrivate::setWindowTitle_sys
would previously already be evaluated to the new
accessible name by
qt_setWindowTitle_helperHelper instead of
using the old window title still set on the
window before the actual title gets updated.

Fixes: QTBUG-128296
Change-Id: Iaf80cd4019b6bc7383f425b45ece51f322a9e1c4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>