QCommandLineParser: Support ignoring input past a certain option
This change adds a new flag to QCommandLineOption to indicate
that any command line options that follow are to be ignored by
QCommandLineParser. When an option with that flag is parsed,
any subsequent options are immediately skipped, and the parser
will return true, indicating success.

This is motivated by Qt WebEngine's need for being able to pass
through command line options from the host application to
the underlying Chromium. This is done by
using the --webEngineArgs option, which separates the Qt application
options (on the left), and the Chromium ones (on the right).
With this change, this mechanism will no longer clash with
QCommandLineParser's behavior of shutting the app down when
encountering an unknown option.

Change-Id: I31aedb7a18e551def742891623604513a329d7ca
Reviewed-by: David Faure <david.faure@kdab.com>