a11y tests: Wait until window is exposed, skip focus test on Wayland
While it passes for QT_QPA_PLATFORM=xcb in a KDE Plasma Dev
Wayland session on Debian testing, tst_QAccessibility::listTest
failed consistently like this for me when run with
QT_QPA_PLATFORM=wayland:

    1: FAIL!  : tst_QAccessibility::listTest() 'QTestAccessibility::containsEvent(&focusEvent)' returned FALSE. ()
    1:    Loc: [.../qt5/qtbase/tests/auto/other/qaccessibility/tst_qaccessibility.cpp(2945)]

The same failure can be observed with QT_QPA_PLATFORM=xcb
when removing the calls to QTest::qWait and
QCoreApplication::processEvents after the call to
QListView::show.

A QAccessible::Focus event is only sent for a newly selected item
when the list view has focus (see QListView::currentChanged),
which was not the case on Wayland.

After the call to QListView::show, wait for the window
to be exposed instead of calling QCoreApplication::processEvents
and waiting for a hard-coded amount of time.

Skip the focus event tests on platforms where focus cannot
be ensured (including Wayland).

Change-Id: I4515a194885b8bac75c6b58a8857ff2152640dd6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>