QLocale: mark c_locale Q_CONSTINIT and c_private() noexcept
Add Q_CONSTINIT to the c_locale variable to statically assert that
it cannot fail (no runtime initialization = no failure). The
QLocalePrivate ctor is already constexpr, so Q_CONSTINIT works out
of the box here.

Because c_locale construction is now known not to fail, c_private()
cannot fail, either, and can and should be noexcept.

Pick-to: 6.8
Change-Id: Ibe94cb4bcd990fbd0b440dd98509eb39447a94b7
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>