QMetaType::id: remove the need to "keep in sync" by simplifying
This was done in 6.1, before we learned better on how to implement these
things. That's also why it uses an int overload instead of
QT6_DECL_NEW_OVERLOAD.

Unfortunately, because it doesn't use the hack, the id(int) overload
isn't declared when compiling removed_api.cpp, so we can't call it.

Change-Id: Idaf6d37063291b019399fffd9af0849633a8ad7f
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>