QMetaType: inline isValid() and isRegistered()
isValid() is (or could be) used in a lot of places, so make it inline.
One could effectively write it as:
    mt == QMetaType()
And rely on comparesEqual() knowing that an invalid type cannot equal to
a valid one.

I'm inlining isRegistered() because it's a simple function.

I chose to use the QT6_DECL_NEW_OVERLOAD technique instead of
QT_CORE_INLINED_SINCE because the latter does not apply inlining to
QtCore itself (aside from LTO).

Change-Id: I0e727ab80c42d51867d0fffd2f9641bc88f926b1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>