QMetaMethod: make some QByteArray-returning methods slightly faster
QByteArray::fromRawData() allocates no memory. Since we know that the
data range is valid, there's no precondition violation either. But we
can only use it for static-lifetime meta objects: those constructed by
QMetaObjectBuilder or QtDBus may get deallocated, causing the strings
obtained from them to crash on use if they do outlive (unlikely, but not
impossible). To differentiate, this commit introduces a new flag to the
QMetaObject header and makes use of it in those two places.

Come Qt 7, we should change these functions to return QByteArrayView,
making the retention of the data in a QByteArray the responsibility of
the user. Making the change right now with #if or QTx_ONLY() is ugly, so
this commit just leaves a comment.

Change-Id: I5aaddbc7ce3fc7a70e15fffd29e276c79d5ef6e4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>